examples/complex_meetings/requirements.txt
openai
instructor
pydantic
typer
pyperclip
rich
pyyaml
